refactor: Extracted context to shared-test-setups
