+++ tests/helpers/shared-test-setups/javascript.js
@@ -5,6 +5,31 @@ const codemodOptions = {
   testAppName: undefined,
 };
 
+const context = {
+  addon: {
+    appReexports: [
+      'components/container-query.js',
+      'helpers/aspect-ratio.js',
+      'helpers/height.js',
+      'helpers/width.js',
+      'modifiers/container-query.js',
+    ],
+    publicEntrypoints: [
+      'components/container-query.js',
+      'helpers/aspect-ratio.js',
+      'helpers/height.js',
+      'helpers/width.js',
+      'modifiers/container-query.js',
+    ],
+  },
+  projectRoot: {
+    devDependencies: {
+      concurrently: '^7.6.0',
+      prettier: '^2.8.1',
+    },
+  },
+};
+
 const options = {
   locations: {
     addon: 'ember-container-query',
@@ -93,4 +118,4 @@ const options = {
   projectRoot: 'tmp/ember-container-query-javascript',
 };
 
-export { codemodOptions, options };
+export { codemodOptions, context, options };
