+++ tests/migration/ember-addon/steps/analyze-addon/javascript.test.js
@@ -1,6 +1,7 @@
 import { analyzeAddon } from '../../../../../src/migration/ember-addon/steps/index.js';
 import {
   codemodOptions,
+  context,
   options,
 } from '../../../../helpers/shared-test-setups/javascript.js';
 import { assert, loadFixture, test } from '../../../../helpers/testing.js';
@@ -43,28 +44,5 @@ test('migration | ember-addon | steps | analyze-addon > javascript', function ()
 
   loadFixture(inputProject, codemodOptions);
 
-  assert.deepEqual(analyzeAddon(options), {
-    addon: {
-      appReexports: [
-        'components/container-query.js',
-        'helpers/aspect-ratio.js',
-        'helpers/height.js',
-        'helpers/width.js',
-        'modifiers/container-query.js',
-      ],
-      publicEntrypoints: [
-        'components/container-query.js',
-        'helpers/aspect-ratio.js',
-        'helpers/height.js',
-        'helpers/width.js',
-        'modifiers/container-query.js',
-      ],
-    },
-    projectRoot: {
-      devDependencies: {
-        concurrently: '^7.6.0',
-        prettier: '^2.8.1',
-      },
-    },
-  });
+  assert.deepEqual(analyzeAddon(options), context);
 });
